chore: move accessSchoolUsers tests into one file
